# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the authentication service used across all sliitmozilla",
        "title": "SLIIT Mozilla Club Auth Service",
        "contact": {},
        "version": "1.0"
    },
    "host": "auth.sliitmozilla.org",
    "paths": {
        "/login": {
            "post": {
                "description": "Endpoint to log in a user with credentials or session token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "Username",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Password",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful, returns session info",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid request or missing parameters",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "description": "Endpoint to log out a user and invalidate their session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Logout user",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized / session not found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    }
}